Make formatting consistant and chage accessor level for some methods
In an effort to remain consistant, the accessor level of BeginPlay()  in the Player and AI controllers has been changed to protected from public, as this is the default level that Unreal sets for a new Pawn class.  In regards to layout in header files, the methods will be defined in order of their accessor level - public, protected, then private.  In addition, a blank line will be kept between the last line of code, and the closing brace in functions; again, this seems to be in keeping with the formatting of the default Unreal boilerplate functions.