Make tank turret rotate towards the point aimed at by the crosshair
This involved creating a new class; TankTurret that inherits from the StaticMeshComponentClass.  The turret on the tank was then removed and recreated using this new class.  The implementation of the movement is very similar as to how the movement of the tank barrel was achieved.  Also, some minor playtesting was done, and as a result, the default values for the movement speeds of the barrel, turret, and projectile were all altered.