Fix bug where AI tanks fly out of control when playing from the main menu
This was fixed by changing the Collision Presets on each tank track in the tank blueprint, to "OverlapAll".  I think the bug was caused when the collision meshes on the tracks and tank body interacted with each other.  I would guess that the only reason it happened from the start menu, is to do with the order in which the assets were loaded when not pre-loaded when playing directly from the scene.  I don't forsee this change affecting anything in the future, as for the scope of this game, collision on the tank body only should suffice.