Move away from using EditAnywhere to EditDefaultsOnly
EditAnywhere allows the property to be changed on specific instances of an object, whereas EditDefaultsOnly only allows changes to be made on in the base class's blueprint.  As we don't necessarily want unique behaviour on all instances of the tanks spawned in the world, changing the properties to EditDefaultsOnly makes more sense, and prevents changes that were not intended.