Change collision meshes on tanks to prevent the tracks' colliding with the body
When the tracks collide with the body, the AI tanks go flying off in all directions when playing the game from the start menu.  Initially, the collision settings on the tracks was set to OverlapAll to prevent this.  In this commit, all collision meshes have been removed, and a single simple box mesh has been added to the tank body.  Not only does this have the same effect as setting the collision meshes on the tank tracks to OverlapAll, it also significantly reduces the complexity of the meshes involved, providing a more elegant solution.