Change collision meshes on the tank blueprint
Previously, the collision meshes had been removed from all components apart from the tank body.  This was to prevent different colliders on the tank's components colliding with one another.  However, as the friction values are going to be worked on next, it makes sense to provide a more realistic collision model that only have the tracks in contact with the ground, rather than using a crude single box collision model which greatly exaggerates the area of contact between the tank and the ground.  In this commit, the tank body collision mesh has been shrunk, and the tank tracks now have their own collision meshes again.  The combination of these meshes gives a good approximation of the tank.