Add UTankMovementComponent class
This class will handle the controls for the tanks, as we introduce a fly-by-wire control scheme.  This new control scheme will work in conjunction with the existing controls, but will allow the AI tanks to effectively control the tanks too.  Because both player and AI tanks will move using the same control system, this should result in a fair game.