Add a TankMovementComponent to the Tank object
Created input bindings for the left thumbstick, and wired them up in the Tank blueprint, so that the left thumbstick (and keyboard "Up") call IntendMoveForward() from the TankMovementComponent, passing in a value of between +1 and -1 (-0.5 for keyboard controls; clamping of -0.5 for gamepad will be done later). Also remove diagnostic log from UTankTrack (no longer needed), and add log to the IntendMovement() method as that is the current area of development.