Fix a bug with AI tanks clipping when game played from the main menu
When the battleground is loaded from the main menu, the AI tanks' left track clips through the terrain.  This seems to be due to a strange order in which the tanks' components are initialised.  Turning the physics for the tank body off in blueprint, and then turning it back on in the events graph after all components of the tank have been initialised, fixes this issue.