Setup code to allow crosshair colour to be changed programatically
Change access modifiers of  GetControlledTank() to allow it to be called from the PlayerController.  Create a blueprint variable to store a reference to the aiming component in the PlayerUI blueprint, then create a binding to the crosshair component that will allow the colour to be changed depending on whether the player tank is ready to fire.