Refactor code to allow Blueprint to handle the initialisation of the UTankAimingComponent
Currenty the code compiles, but does not run as expected - all Tank aiming and firing functionality is broken.  This is due to the TankAimingComponent not being correctly found as the architecture is being shifted.  I hate committing broken code, but completeing the architecture transition is a bigger task than I thought, and as I'm moving in the right direction, I think at this point, it's worth making a commit.

Also change all Blueprint categories to use quotation marks to avoid potential compilation issues.