Add an event driven paradigm for creating the Player UI component
This fixes the UI issue where the crosshair wasn't changing colour because the UTankPlayerController couldn't find a UTankAimingComponent.  Now, the component is found in the BeginPlay() method of UTankPlayerController, and triggers an event to find the Aiming Component.  Because the aiming component is already constructed when BeginPlay() is called, should always find this component.  As the UTankAimingComponent is now correctly wired up, the crosshair now correctly changes colour depending on what is set in the TankAimingComponent.cpp file.  This should automatically update once the aiming component's functionality is fixed.