Migrate pointer protection to utilise the "ensure" macro
Before returning from a function containing a nullptr, the ensure logs out where the nullptr occurred, making this great for debugging issues.  From this, it's clear that to get the code to work, I am missing a TankAimingComponent and a Barrel in the Tank.cpp files.  This makes sense, as I removed these components when shifting the architecture of the ATank class.