Refactor AimAt() out of ATank
The goal is to completely remove any dependencies from tank and to streamline the architecture.  This first step is to remove the handler function AimAt() from Tank, and let the AI and Player controllers directly call the actual AimAt() function from the Tank Aiming Component.  This also allows several dependencies to be broken, which creates a simpler program structure.