Log to screen when a tank track is in contact with the ground
This is setup work to facilitate a force only being applied when the tank is on the ground.  Currently, a force can still be applied when the tank is in the air which can make the tank move erratically through the air.  Applying a force only when the tank is grounded should prevent this from happening.  Also replaced deprecating GetRootPrimitiveComponent().