Enable mortars to be "destroyed"
Setting physics with a low weight on the mortar body allows it to be hit, and flung a long way on impact.  This makes the mortar model travel outside the world bounds, and gets destroyed by the engine.  This is not ideal, but loosely simulates the ability to destroy the mortar.