Fix bug where ensure was triggering on player death
On player death, the player controller depossesses the tank and enters spectator mode.  This causes the ensure to fail in the TankAIController class.  Moving the PlayerTank setter to BeginPlay() would solve the issue, but if later in development the ability for the player to respawn is introduced, then this setter will have to be redone anyway.

Fixes #5